Keep earliest start time per domain
This is important on OCaml 4 where we use domain-shims, which uses systhreads to
emulate domains.  What can happen with systhreads, in particular, is that one or
more systhreads may run for a long time or even complete their work before other
systhreads get a change to run.  The timings captured on different workers are
then (more or less) sequential rather than parallel as assumed by the
calculation of time per operation (latency) and total thruput.