Randomize the order of benchmark suites by default
The intention behind this is to try to account for things that might put the
runtime into a state where it has major effects on specific benchmarks.  After
randomizing, we should see variation from run to run rather than e.g. having the
potential effects only shown in specific suites.